Avoid poseball sit being overridden by AO sit
While we where sat on a poseball, the AO overrides the new pose for
Walking in the timer which caused our current sit pose to revert to our
overridden sit pose instead of the animation that the poseball plays. A
bit hard to explain sorry. This might be an OpenSim specific bug.